// models/paginate.ts
export interface Paginate<T> {
  content: T[];
  pageable: Pageable;
  totalPages: number;
  totalElements: number;
  last?: boolean;
  first?: boolean;
  sort?: Sort;
  numberOfElements?: number;
  size?: number;
  number?: number;
  empty?: boolean;
}

export interface Pageable {
  sort: Sort;
  pageNumber: number;
  pageSize: number;
  offset: number;
  unpaged: boolean;
  paged: boolean;
}

export interface Sort {
  sorted: boolean;
  unsorted: boolean;
  empty: boolean;
}